// src/routes/vehicles-management/enums/customers-vehicle.enum.ts
export enum EVehicleStatus {
  INSPECTING = 'INSPECTING',
  PENDING = 'PENDING',
  WAITING_PARTS = 'WAITING_PARTS',
  REPAIRING = 'REPAIRING',
  QUALITY_CHECK = 'QUALITY_CHECK',
  WAITING_APPROVAL = 'WAITING_APPROVAL',
  READY_FOR_PICKUP = 'READY_FOR_PICKUP',
  COMPLETED = 'COMPLETED',
  CANCELLED = 'CANCELLED',
}